lung-cancer-classification/requirements.txt
# ============================================================================
# LUNG CANCER CLASSIFICATION PROJECT - DEPENDENCIES
# Optimized for Windows 11, Intel i5 13th Gen, RTX 2050 4GB
# Python 3.13.1
# ============================================================================

# Deep Learning Framework
tensorflow==2.20.0              # Deep learning library with GPU support (Python 3.13 compatible)

# Data Processing & Manipulation
numpy==1.26.2                   # Numerical computing
pandas==2.1.4                   # Data manipulation and analysis
pillow==10.1.0                  # Image processing library
opencv-python==4.8.1.78         # Computer vision operations
scikit-image==0.22.0            # Image processing algorithms

# Machine Learning Utilities
scikit-learn==1.3.2             # ML algorithms and metrics
scipy==1.11.4                   # Scientific computing

# Visualization
matplotlib==3.8.2               # Plotting and visualization
seaborn==0.13.0                 # Statistical data visualization
plotly==5.18.0                  # Interactive plots

# Model Interpretability
tf-keras-vis==0.8.5             # GradCAM and model visualization
lime==0.2.0.1                   # Model explanation

# Progress Bars & Utilities
tqdm==4.66.1                    # Progress bars
pyyaml==6.0.1                   # Configuration file handling
ipywidgets==8.1.1               # Jupyter widgets

# Jupyter Notebook Support
jupyter==1.0.0                  # Jupyter notebook
notebook==7.0.6                 # Jupyter notebook interface
ipykernel==6.27.1               # Jupyter kernel

# GPU Memory Management (Important for 4GB VRAM)
gputil==1.4.0                   # GPU monitoring

# Additional Utilities
colorama==0.4.6                 # Colored terminal output (Windows support)
python-dotenv==1.0.0            # Environment variable management

# Optional but Recommended
tensorboard==2.15.1             # Training visualization
